+++ models/exclusions.js
@@ -18,7 +18,7 @@ const Exclusions = sequelize.define('exclusions', {
         type: Sequelize.DATEONLY,
         allowNull: false
     },
-    addAmt: { type: Sequelize.DECIMAL(10, 2) },
+    addAmt: { type: Sequelize.FLOAT(10, 2) },
     addMemo: { type: Sequelize.STRING },
 
 });
